+++ Board/src/main/java/com/board/Repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.board.DAO;
+package com.board.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.board.Entity.Member;
 
 @Repository
-public interface MemberDAO extends JpaRepository<Member, String> {
+public interface MemberRepository extends JpaRepository<Member, String> {
 
 }
